+++ src/ebay-chip/index.ts
@@ -1,0 +1,1 @@
+export { default as EbayChip, type EbayChipProps } from './chip'