+++ src/ebay-dialog-base/components/dialogBase.tsx
@@ -7,10 +7,8 @@ import React, {
     useState,
     ReactElement,
     cloneElement,
-    MouseEventHandler,
     ReactNode,
-    KeyboardEvent,
-    KeyboardEventHandler
+    KeyboardEvent
 } from 'react'
 import classNames from 'classnames'
 import * as screenreaderTrap from 'makeup-screenreader-trap'
@@ -18,6 +16,7 @@ import * as keyboardTrap from 'makeup-keyboard-trap'
 import { EbayIcon } from '../../ebay-icon'
 import { randomId } from '../../common/random-id'
 import { useDialogAnimation, TransitionElement } from './animation'
+import { DialogCloseEventHandler } from '../types'
 
 
 export type WindowType = 'compact'
@@ -42,8 +41,8 @@ export interface DialogBaseProps<T> extends HTMLProps<T> {
     ariaLabelledby?: string;
     a11yCloseText?: string;
     onOpen?: () => void;
-    onCloseBtnClick?: MouseEventHandler & KeyboardEventHandler;
-    onBackgroundClick?: MouseEventHandler;
+    onCloseBtnClick?: DialogCloseEventHandler;
+    onBackgroundClick?: DialogCloseEventHandler;
     mainId?: string;
     ignoreEscape?: boolean;
     closeButton?: ReactElement;
@@ -97,7 +96,7 @@ export const DialogBase: FC<DialogBaseProps<HTMLElement>> = ({
 
     useEffect(() => {
         let timeout: number
-        const handleBackgroundClick = (e: React.MouseEvent) => {
+        const handleBackgroundClick = (e: React.MouseEvent<HTMLElement>) => {
             if (drawerBaseEl.current && !drawerBaseEl.current.contains(e.target)) {
                 onBackgroundClick(e)
             }
@@ -141,7 +140,7 @@ export const DialogBase: FC<DialogBaseProps<HTMLElement>> = ({
         onTransitionEnd: () => handleFocus(open)
     })
 
-    const onKeyDown = (event: KeyboardEvent) => {
+    const onKeyDown = (event: KeyboardEvent<HTMLElement>) => {
         if (!ignoreEscape && event.key === 'Escape') {
             event.stopPropagation()
             onCloseBtnClick(event)